+++ contracts/src/main/kotlin/com/r3corda/contracts/CommercialPaper.kt
@@ -1,5 +1,8 @@
 package com.r3corda.contracts
 
+import com.r3corda.contracts.cash.Cash
+import com.r3corda.contracts.cash.InsufficientBalanceException
+import com.r3corda.contracts.cash.sumCashBy
 import com.r3corda.core.*
 import com.r3corda.core.contracts.*
 import com.r3corda.core.crypto.NullPublicKey