+++ contracts/src/test/kotlin/com/r3corda/contracts/CrowdFundTests.kt
@@ -1,5 +1,6 @@
 package com.r3corda.contracts
 
+import com.r3corda.contracts.cash.Cash
 import com.r3corda.contracts.testing.CASH
 import com.r3corda.contracts.testing.`owned by`
 import com.r3corda.core.contracts.*